pyproject.toml
[project]
name = "retro-asset-gen"
version = "1.0.0"
description = "Generate retro gaming platform assets using Gemini AI for Pegasus Frontend themes"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
]

[project.scripts]
retro-asset-gen = "retro_asset_gen.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/retro_asset_gen"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/retro_asset_gen/prompts.py" = ["E501"]  # Long prompt strings are intentional

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
